move yaml to conda.common.yaml
