+++ conda/utils.py
@@ -277,32 +277,6 @@ def find_parent_shell(path=False, max_stack_depth=10):
     return process.parent().name()
 
 
-@memoized
-def get_yaml():
-    try:
-        import ruamel_yaml as yaml
-    except ImportError:                                         # pragma: no cover
-        try:                                                    # pragma: no cover
-            import ruamel.yaml as yaml                          # pragma: no cover
-        except ImportError:                                     # pragma: no cover
-            raise ImportError("No yaml library available.\n"    # pragma: no cover
-                              "To proceed, conda install "      # pragma: no cover
-                              "ruamel_yaml")                    # pragma: no cover
-    return yaml
-
-
-def yaml_load(filehandle):
-    yaml = get_yaml()
-    return yaml.load(filehandle, Loader=yaml.RoundTripLoader, version="1.1")
-
-
-def yaml_dump(string):
-    yaml = get_yaml()
-    return yaml.dump(string, Dumper=yaml.RoundTripDumper,
-                     block_seq_indent=2, default_flow_style=False,
-                     indent=4)
-
-
 # TODO: this should be done in a more extensible way
 #     (like files for each shell, with some registration mechanism.)
